+++ app/upgrades/v8/incentive_props.go
@@ -2,9 +2,9 @@ package v8
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/osmosis-labs/osmosis/v11/osmoutils"
-	poolincentiveskeeper "github.com/osmosis-labs/osmosis/v11/x/pool-incentives/keeper"
-	poolincentivestypes "github.com/osmosis-labs/osmosis/v11/x/pool-incentives/types"
+	"github.com/osmosis-labs/osmosis/v12/osmoutils"
+	poolincentiveskeeper "github.com/osmosis-labs/osmosis/v12/x/pool-incentives/keeper"
+	poolincentivestypes "github.com/osmosis-labs/osmosis/v12/x/pool-incentives/types"
 )
 
 // This file implements logic for accelerated incentive proposals.