+++ x/epochs/types/hooks.go
@@ -1,16 +1,18 @@
 package types
 
 import (
+	fmt "fmt"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/osmosis-labs/osmosis/v11/osmoutils"
+	"github.com/osmosis-labs/osmosis/v12/osmoutils"
 )
 
 type EpochHooks interface {
 	// the first block whose timestamp is after the duration is counted as the end of the epoch
-	AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64)
+	AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) error
 	// new epoch is next block of epoch end block
-	BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64)
+	BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64) error
 }
 
 var _ EpochHooks = MultiEpochHooks{}
@@ -23,32 +25,33 @@ func NewMultiEpochHooks(hooks ...EpochHooks) MultiEpochHooks {
 }
 
 // AfterEpochEnd is called when epoch is going to be ended, epochNumber is the number of epoch that is ending.
-func (h MultiEpochHooks) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) {
+func (h MultiEpochHooks) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) error {
 	for i := range h {
 		panicCatchingEpochHook(ctx, h[i].AfterEpochEnd, epochIdentifier, epochNumber)
 	}
+	return nil
 }
 
 // BeforeEpochStart is called when epoch is going to be started, epochNumber is the number of epoch that is starting.
-func (h MultiEpochHooks) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64) {
+func (h MultiEpochHooks) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64) error {
 	for i := range h {
 		panicCatchingEpochHook(ctx, h[i].BeforeEpochStart, epochIdentifier, epochNumber)
 	}
+	return nil
 }
 
 func panicCatchingEpochHook(
 	ctx sdk.Context,
-	hookFn func(ctx sdk.Context, epochIdentifier string, epochNumber int64),
+	hookFn func(ctx sdk.Context, epochIdentifier string, epochNumber int64) error,
 	epochIdentifier string,
 	epochNumber int64,
 ) {
-	defer func() {
-		if recovErr := recover(); recovErr != nil {
-			osmoutils.PrintPanicRecoveryError(ctx, recovErr)
-		}
-	}()
-	cacheCtx, write := ctx.CacheContext()
-	hookFn(cacheCtx, epochIdentifier, epochNumber)
-	write()
-	ctx.EventManager().EmitEvents(cacheCtx.EventManager().Events())
+	wrappedHookFn := func(ctx sdk.Context) error {
+		return hookFn(ctx, epochIdentifier, epochNumber)
+	}
+	// TODO: Thread info for which hook this is, may be dependent on larger hook system refactoring
+	err := osmoutils.ApplyFuncIfNoError(ctx, wrappedHookFn)
+	if err != nil {
+		ctx.Logger().Error(fmt.Sprintf("error in epoch hook %v", err))
+	}
 }