+++ x/pool-incentives/module.go
@@ -4,9 +4,9 @@ in the `incentives` module for every lock duration
 that exists in that pool. The `pool-incentives` module also takes
 the `pool_incentives` distributed from the `gov` module
 and distributes it to the various incentivized gauges.
- - Handles governance proposals impacting pool incentives.
- - Pool distribution and lockup infos queries.
- - Distributes incentives to LPs.
+  - Handles governance proposals impacting pool incentives.
+  - Pool distribution and lockup infos queries.
+  - Distributes incentives to LPs.
 */
 package pool_incentives
 
@@ -29,9 +29,9 @@ import (
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	"github.com/cosmos/cosmos-sdk/x/gov/simulation"
 
-	"github.com/osmosis-labs/osmosis/v11/x/pool-incentives/client/cli"
-	"github.com/osmosis-labs/osmosis/v11/x/pool-incentives/keeper"
-	"github.com/osmosis-labs/osmosis/v11/x/pool-incentives/types"
+	"github.com/osmosis-labs/osmosis/v12/x/pool-incentives/client/cli"
+	"github.com/osmosis-labs/osmosis/v12/x/pool-incentives/keeper"
+	"github.com/osmosis-labs/osmosis/v12/x/pool-incentives/types"
 )
 
 var (
@@ -67,7 +67,7 @@ func (AppModuleBasic) ValidateGenesis(cdc codec.JSONCodec, config client.TxEncod
 	return types.ValidateGenesis(&data)
 }
 
-//---------------------------------------
+// ---------------------------------------
 // Interfaces.
 func (b AppModuleBasic) RegisterRESTRoutes(ctx client.Context, r *mux.Router) {
 	// noop