+++ x/twap/types/genesis_test.go
@@ -87,7 +87,7 @@ func TestGenesisState_Validate(t *testing.T) {
 		"valid empty records": {
 			twapGenesis: NewGenesisState(basicParams, []TwapRecord{}),
 		},
-		"invalid genesis - error": {
+		"invalid genesis - pool ID doesn't exist": {
 			twapGenesis: NewGenesisState(
 				NewParams("week", 48*time.Hour),
 				[]TwapRecord{
@@ -145,11 +145,11 @@ func TestGenesisState_Validate(t *testing.T) {
 }
 
 func TestTWAPRecord_Validate(t *testing.T) {
-	testCases := map[string]struct {
-		twapRecord TwapRecord
-
+	type testcase struct {
+		twapRecord  TwapRecord
 		expectedErr bool
-	}{
+	}
+	testCases := map[string]testcase{
 		"valid base record": {
 			twapRecord: baseRecord,
 		},
@@ -162,19 +162,10 @@ func TestTWAPRecord_Validate(t *testing.T) {
 
 			expectedErr: true,
 		},
-		"invalid asset0 denom": {
+		"invalid denom": {
 			twapRecord: func() TwapRecord {
 				r := baseRecord
 				r.Asset0Denom = ""
-				return r
-			}(),
-
-			expectedErr: true,
-		},
-		"invalid asset1 denom": {
-			twapRecord: func() TwapRecord {
-				r := baseRecord
-				r.Asset1Denom = ""
 				return r
 			}(),
 
@@ -198,7 +189,7 @@ func TestTWAPRecord_Validate(t *testing.T) {
 
 			expectedErr: true,
 		},
-		"invalid p0 last spot price": {
+		"invalid p0 last spot price: zero": {
 			twapRecord: func() TwapRecord {
 				r := baseRecord
 				r.P0LastSpotPrice = sdk.ZeroDec()
@@ -207,16 +198,61 @@ func TestTWAPRecord_Validate(t *testing.T) {
 
 			expectedErr: true,
 		},
-		"invalid p1 last spot price": {
+		"invalid p0 last spot price: negative": {
 			twapRecord: func() TwapRecord {
 				r := baseRecord
-				r.P1LastSpotPrice = sdk.ZeroDec()
+				r.P0LastSpotPrice = sdk.OneDec().Neg()
 				return r
 			}(),
 
 			expectedErr: true,
 		},
-		"invalid p0 arithmetic accum": {
+		"invalid last spot price with error": {
+			twapRecord: func() TwapRecord {
+				r := baseRecord
+				r.LastErrorTime = r.Time
+				r.P0LastSpotPrice = sdk.NewDec(5) // has to be distinct to be symmetric
+				r.P1LastSpotPrice = sdk.ZeroDec() // need this to be 0, to test the other case on error
+				return r
+			}(),
+
+			expectedErr: true,
+		},
+		"invalid p0 last spot price: nil": {
+			twapRecord: func() TwapRecord {
+				r := TwapRecord{
+					PoolId:                      basePoolId,
+					Asset0Denom:                 denom0,
+					Asset1Denom:                 denom1,
+					Height:                      3,
+					Time:                        tPlusOne.Add(time.Second),
+					P1LastSpotPrice:             sdk.OneDec(),
+					P0ArithmeticTwapAccumulator: sdk.OneDec(),
+					P1ArithmeticTwapAccumulator: sdk.OneDec(),
+				}
+				return r
+			}(),
+
+			expectedErr: true,
+		},
+		"invalid p1 last spot price: nil": {
+			twapRecord: func() TwapRecord {
+				r := TwapRecord{
+					PoolId:                      basePoolId,
+					Asset0Denom:                 denom0,
+					Asset1Denom:                 denom1,
+					Height:                      3,
+					Time:                        tPlusOne.Add(time.Second),
+					P0LastSpotPrice:             sdk.OneDec(),
+					P0ArithmeticTwapAccumulator: sdk.OneDec(),
+					P1ArithmeticTwapAccumulator: sdk.OneDec(),
+				}
+				return r
+			}(),
+
+			expectedErr: true,
+		},
+		"invalid p0 arithmetic accum: negative": {
 			twapRecord: func() TwapRecord {
 				r := baseRecord
 				r.P0ArithmeticTwapAccumulator = sdk.OneDec().Neg()
@@ -225,18 +261,63 @@ func TestTWAPRecord_Validate(t *testing.T) {
 
 			expectedErr: true,
 		},
-		"invalid p1 arithmetic accum": {
+		"invalid p0 arithmetic accum: nil": {
 			twapRecord: func() TwapRecord {
-				r := baseRecord
-				r.P1ArithmeticTwapAccumulator = sdk.OneDec().Neg()
+				r := TwapRecord{
+					PoolId:                      basePoolId,
+					Asset0Denom:                 denom0,
+					Asset1Denom:                 denom1,
+					Height:                      3,
+					Time:                        tPlusOne.Add(time.Second),
+					P0LastSpotPrice:             sdk.OneDec(),
+					P1LastSpotPrice:             sdk.OneDec(),
+					P1ArithmeticTwapAccumulator: sdk.OneDec(),
+				}
 				return r
 			}(),
 
 			expectedErr: true,
 		},
+		"invalid p1 arithmetic accum: nil": {
+			twapRecord: func() TwapRecord {
+				r := TwapRecord{
+					PoolId:                      basePoolId,
+					Asset0Denom:                 denom0,
+					Asset1Denom:                 denom1,
+					Height:                      3,
+					Time:                        tPlusOne.Add(time.Second),
+					P0LastSpotPrice:             sdk.OneDec(),
+					P1LastSpotPrice:             sdk.OneDec(),
+					P0ArithmeticTwapAccumulator: sdk.OneDec(),
+				}
+				return r
+			}(),
+			expectedErr: true,
+		},
+	}
+	// make test cases symmetric
+	testCasesSym := map[string]testcase{}
+	for k, tc := range testCases {
+		if tc.twapRecord.P0LastSpotPrice.IsNil() ||
+			tc.twapRecord.P1LastSpotPrice.IsNil() ||
+			tc.twapRecord.P0ArithmeticTwapAccumulator.IsNil() ||
+			tc.twapRecord.P1ArithmeticTwapAccumulator.IsNil() {
+			testCasesSym[k] = tc
+		} else if tc.twapRecord.Asset0Denom != baseRecord.Asset0Denom ||
+			!tc.twapRecord.P0LastSpotPrice.Equal(baseRecord.P0LastSpotPrice) ||
+			!tc.twapRecord.P0ArithmeticTwapAccumulator.Equal(baseRecord.P0ArithmeticTwapAccumulator) {
+			testCasesSym[k+": asset 0"] = tc
+			tSym := tc.twapRecord
+			tSym.Asset0Denom, tSym.Asset1Denom = tSym.Asset1Denom, tSym.Asset0Denom
+			tSym.P0LastSpotPrice, tSym.P1LastSpotPrice = tSym.P1LastSpotPrice, tSym.P0LastSpotPrice
+			tSym.P0ArithmeticTwapAccumulator, tSym.P1ArithmeticTwapAccumulator = tSym.P1ArithmeticTwapAccumulator, tSym.P0ArithmeticTwapAccumulator
+			testCasesSym[k+": asset 1"] = testcase{tSym, tc.expectedErr}
+		} else {
+			testCasesSym[k] = tc
+		}
 	}
 
-	for name, tc := range testCases {
+	for name, tc := range testCasesSym {
 		t.Run(name, func(t *testing.T) {
 			// System under test.
 			err := tc.twapRecord.Validate()